+++ packages/io/src/streams/index.ts
@@ -12,7 +12,12 @@ import { Readable, type ReadableState, addAbortSignal } from './readable.ts';
 import { Transform } from './transform.ts';
 import { Writable, type WritableState } from './writable.ts';
 
-export { Readable, type ReadableFromWebOptions, type ReadableOptions } from './readable.ts';
+export {
+  Readable,
+  type ReadableFromWebOptions,
+  type ReadableOptions,
+  type AsyncHelperOptions,
+} from './readable.ts';
 export { Writable, type WritableOptions, type WriteChunk } from './writable.ts';
 export { Duplex, type DuplexOptions } from './duplex.ts';
 export { Transform, type TransformOptions } from './transform.ts';